// packages/silk/src/skin.generated.ts
export const SKIN = {
// ...
  "ActionPanel": {
    "x": 486.52,
    "y": 10,
    "w": 348.69,
    "h": 70,
    "bg": "rgb(38, 38, 38)",
    "color": "rgb(250, 250, 250)",
    "radius": 3,
    "font": {
      "size": 16,
      "weight": "400",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 3,
      "color": "rgb(13, 13, 13)"
    },
    "shadow": null,
    "text": "BET$1.000"
  },
  "FeatureBuyToggle": {
    "x": -118.2,
    "y": -14.4,
    "w": 118.79,
    "h": 118.79,
    "bg": "rgb(255, 198, 55)",
    "color": "rgb(0, 0, 0)",
    "radius": 100,
    "font": {
      "size": 14,
      "weight": "700",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "capitalize"
    },
    "border": {
      "width": 0,
      "color": "rgb(0, 0, 0)"
    },
    "shadow": null,
    "text": "BUY BONUS"
  },
  "MainMenuToggle": {
    "x": 5.8,
    "y": 17,
    "w": 55,
    "h": 56,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(250, 250, 250)",
    "radius": 0,
    "font": {
      "size": 26,
      "weight": "500",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "capitalize"
    },
    "border": {
      "width": 0,
      "color": "rgb(250, 250, 250)"
    },
    "shadow": null,
    "text": ""
  },
  "BalanceItem": {
    "x": 73.3,
    "y": 26.39,
    "w": 113.7,
    "h": 37.22,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(250, 250, 250)",
    "radius": 0,
    "font": {
      "size": 16,
      "weight": "400",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(250, 250, 250)"
    },
    "shadow": null,
    "text": "BALANCE$12,345.67"
  },
  "BalanceLabel": {
    "x": 81.3,
    "y": 26.39,
    "w": 65.36,
    "h": 16.09,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(255, 255, 255)",
    "radius": 0,
    "font": {
      "size": 14,
      "weight": "400",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(255, 255, 255)"
    },
    "shadow": null,
    "text": "BALANCE"
  },
  "BalanceValue": {
    "x": 81.3,
    "y": 42.48,
    "w": 105.7,
    "h": 21.13,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(255, 255, 255)",
    "radius": 0,
    "font": {
      "size": 21.12,
      "weight": "700",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(255, 255, 255)"
    },
    "shadow": null,
    "text": "$12,345.67"
  },
  "WinAmountItem": {
    "x": 203,
    "y": 24.95,
    "w": 60.06,
    "h": 40.09,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(250, 250, 250)",
    "radius": 0,
    "font": {
      "size": 16,
      "weight": "400",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(250, 250, 250)"
    },
    "shadow": null,
    "text": "WIN$0.00"
  },
  "WinAmountLabel": {
    "x": 203,
    "y": 24.95,
    "w": 27.22,
    "h": 16.09,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(255, 255, 255)",
    "radius": 0,
    "font": {
      "size": 14,
      "weight": "400",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(255, 255, 255)"
    },
    "shadow": null,
    "text": "WIN"
  },
  "WinAmountValue": {
    "x": 203,
    "y": 41.05,
    "w": 60.06,
    "h": 24,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(255, 255, 255)",
    "radius": 0,
    "font": {
      "size": 24,
      "weight": "700",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(255, 255, 255)"
    },
    "shadow": null,
    "text": "$0.00"
  },
  "BetAmountItem": {
    "x": 489.52,
    "y": 13,
    "w": 120,
    "h": 64,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(250, 250, 250)",
    "radius": 0,
    "font": {
      "size": 16,
      "weight": "400",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(250, 250, 250)"
    },
    "shadow": null,
    "text": "BET$1.00"
  },
  "BetAmountLabel": {
    "x": 504.52,
    "y": 24.95,
    "w": 60.06,
    "h": 16.09,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(255, 255, 255)",
    "radius": 0,
    "font": {
      "size": 14,
      "weight": "400",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(255, 255, 255)"
    },
    "shadow": null,
    "text": "BET"
  },
  "BetAmountValue": {
    "x": 504.52,
    "y": 41.05,
    "w": 60.06,
    "h": 24,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(255, 255, 255)",
    "radius": 0,
    "font": {
      "size": 24,
      "weight": "700",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(255, 255, 255)"
    },
    "shadow": null,
    "text": "$1.00"
  },
  "BetAmountIndicatorProgress": {
    "x": 489.52,
    "y": 72,
    "w": 48,
    "h": 5,
    "bg": "rgb(255, 255, 255)",
    "color": "rgb(250, 250, 250)",
    "radius": 0,
    "font": {
      "size": 16,
      "weight": "400",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(250, 250, 250)"
    },
    "shadow": null,
    "text": ""
  },
  "BetAmountIncrease": {
    "x": 611.09,
    "y": 13,
    "w": 49.02,
    "h": 32,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(250, 250, 250)",
    "radius": 4,
    "font": {
      "size": 12,
      "weight": "500",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "capitalize"
    },
    "border": {
      "width": 0,
      "color": "rgb(250, 250, 250)"
    },
    "shadow": null,
    "text": ""
  },
  "BetAmountDecrease": {
    "x": 611.02,
    "y": 45,
    "w": 49.19,
    "h": 32,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(250, 250, 250)",
    "radius": 4,
    "font": {
      "size": 12,
      "weight": "500",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "capitalize"
    },
    "border": {
      "width": 0,
      "color": "rgb(250, 250, 250)"
    },
    "shadow": null,
    "text": ""
  },
  "PlaceBetBtn": {
    "x": 660.2,
    "y": -11,
    "w": 112,
    "h": 112,
    "bg": "rgb(38, 38, 38)",
    "color": "rgb(250, 250, 250)",
    "radius": 100,
    "font": {
      "size": 86.4,
      "weight": "500",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "capitalize"
    },
    "border": {
      "width": 3,
      "color": "rgb(0, 0, 0)"
    },
    "shadow": null,
    "text": ""
  },
  "AutoplayBtn": {
    "x": 772.2,
    "y": 15,
    "w": 60,
    "h": 60,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(250, 250, 250)",
    "radius": 100,
    "font": {
      "size": 34,
      "weight": "500",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "capitalize"
    },
    "border": {
      "width": 0,
      "color": "rgb(250, 250, 250)"
    },
    "shadow": null,
    "text": ""
  },
  "FeedbackMsg": {
    "x": 0,
    "y": -30,
    "w": 840,
    "h": 20,
    "bg": "rgba(0, 0, 0, 0)",
    "color": "rgb(255, 255, 255)",
    "radius": 0,
    "font": {
      "size": 16,
      "weight": "700",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(255, 255, 255)"
    },
    "shadow": null,
    "text": "PRESS PLAY TO SPIN!"
  },
  "BetAmountProgressbar": {
    "x": 489.52,
    "y": 72,
    "w": 120,
    "h": 5,
    "bg": "rgba(0, 0, 0, 0.9)",
    "color": "rgb(250, 250, 250)",
    "radius": 0,
    "font": {
      "size": 16,
      "weight": "400",
      "family": "HacksawUI",
      "spacing": "normal",
      "transform": "none"
    },
    "border": {
      "width": 0,
      "color": "rgb(250, 250, 250)"
    },
    "shadow": null,
    "text": ""
  },
// ...
} as const;